# lib/api-zod/src/generated/types/generateReportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateReportRequestType } from "./generateReportRequestType";
import type { GenerateReportSpecies } from "./generateReportSpecies";

export interface GenerateReportRequest {
  type: GenerateReportRequestType;
  city: string;
  radiusKm: number;
  group: string;
  recentObservations: number;
  historicalObservations: number;
  uniqueSpecies: number;
  focusSpecies?: GenerateReportSpecies;
  topSpecies: GenerateReportSpecies[];
}