# lib/api-zod/src/generated/types/generateReportResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateReportResponse {
  title: string;
  executiveSummary: string;
  keyFinding: string;
  whyItMatters: string;
  bullets: string[];
  recommendations: string[];
}