# lib/api-zod/src/generated/types/generateReportSpecies.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateReportSpecies {
  commonName: string;
  scientificName: string;
  group?: string;
  role?: string;
  recentCount: number;
  conservation?: string;
}